+++ CHANGELOG.rst
@@ -5,9 +5,10 @@ Changelog
 TBR
 ------------------
 
+* added a ``PageObjectRegistry`` class which has the  ``handle_urls`` decorator
+  to write override rules.
 * removed support for Python 3.6
 * added support for Python 3.10
-
 
 0.1.1 (2021-06-02)
 ------------------